Make settings more consistent
The goal is to provide more sensible defaults and reduce the number of switches necessary
to transition from demo to live and back.

Now the global demo mode enacted by `ESTEID_DEMO` (on by default) also selects other values
as suitable for demo modes, and flipping it to off selects those values for live.

* MobileID and SmartID test mode now defaults to the global demo mode set by ESTEID_DEMO;
* Whether to allow multiple signatures also defaults to ESTEID_DEMO.